2 fix.
- Fix last turtle install in mass install.
- fixed the setting page for filtered storage was 4, now it's 3.

+++ disk_startup.lua
@@ -40,7 +40,7 @@ turtle.select( 1 )
 if turtle.suck( 1 ) or turtle.inspectUp() then
   -- make sure it's a turtle.
   local item = turtle.getItemDetail()
-  if not string.find( item.name, "turtle" ) then
+  if item and not string.find( item.name, "turtle" ) then
     turtle.drop()
     return
   end
